secure todos

+++ functions/APIs/todos.js
@@ -4,6 +4,7 @@ const { db } = require('../util/admin');
 
 exports.getAllTodos = (request, response) => {
 	db.collection('todos')
+		.where('username', '==', request.user.username)
 		.orderBy('createdAt', 'desc')
 		.get()
 		.then(data => {
@@ -36,6 +37,7 @@ exports.postOneTodo = (request, response) => {
 	const newTodoItem = {
 		title: request.body.title,
 		body: request.body.body,
+		username: request.user.username,
 		createdAt: new Date().toISOString()
 	};
 
@@ -59,6 +61,9 @@ exports.deleteTodo = (request, response) => {
 		.then(doc => {
 			if (!doc.exists) {
 				return response.status(404).json({ error: 'Todo not found' });
+			}
+			if (doc.data().username !== request.user.username) {
+				return response.status(403).json({ error: 'UnAuthorized' });
 			}
 			return document.delete();
 		})
